Update.
2003-04-08  Ulrich Drepper  <drepper@redhat.com>

	* tst-barrier2.c: Eliminate warnings.
	* tst-cancel4.c: Likewise.
	* tst-cond4.c: Likewise.
	* tst-cond6.c: Likewise.
	* tst-detach1.c: Likewise.
	* tst-rwlock4.c: Likewise.
	* tst-rwlock6.c: Likewise.
	* tst-rwlock7.c: Likewise.
	* tst-sem3.c: Likewise.
	* tst-spin2.c: Likewise.
	* tst-umask1.c: Likewise.

+++ nptl/tst-barrier2.c
@@ -64,7 +64,7 @@ do_test (void)
   memset (data, '\0', ps);
 
   /* Write the data to the file.  */
-  if (write (fd, data, ps) != ps)
+  if (write (fd, data, ps) != (ssize_t) ps)
     {
       puts ("short write");
       return 1;
